Move logic to SuppressWarningsUtils
